requirements.txt
flask
requests
gunicorn
python-dotenv
openai
redis